+++ baybe/surrogates/gaussian_process/components/__init__.py
@@ -1,5 +1,10 @@
 """Gaussian process surrogate components."""
 
+from baybe.surrogates.gaussian_process.components.fit_criterion import (
+    FitCriterion,
+    FitCriterionFactoryProtocol,
+    PlainFitCriterionFactory,
+)
 from baybe.surrogates.gaussian_process.components.kernel import (
     KernelFactoryProtocol,
     PlainKernelFactory,
@@ -15,6 +20,10 @@ from baybe.surrogates.gaussian_process.components.mean import (
 )
 
 __all__ = [
+    # Fit Criterion
+    "FitCriterion",
+    "FitCriterionFactoryProtocol",
+    "PlainFitCriterionFactory",
     # Kernel
     "KernelFactoryProtocol",
     "PlainKernelFactory",